+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/TxOut.hs
@@ -108,7 +108,7 @@ import Cardano.Ledger.Plutus.Data (
  )
 import Cardano.Ledger.Val (Val (..))
 import Control.DeepSeq (NFData (rnf), rwhnf)
-import Data.Aeson (ToJSON (..), (.=))
+import Data.Aeson (FromJSON (..), ToJSON (..), withObject, (.:), (.=))
 import qualified Data.ByteString.Lazy as LBS
 import Data.Maybe (fromMaybe)
 import Data.MemPack
@@ -331,6 +331,22 @@ instance (Era era, Val (Value era), ToJSON (Script era)) => ToKeyValuePairs (Bab
     , "datum" .= dat
     , "referenceScript" .= mRefScript
     ]
+
+instance
+  ( Era era
+  , Val (Value era)
+  , FromJSON (Value era)
+  , FromJSON (Script era)
+  , FromJSON (Datum era)
+  ) =>
+  FromJSON (BabbageTxOut era)
+  where
+  parseJSON = withObject "BabbageTxOut" $ \o ->
+    BabbageTxOut
+      <$> o .: "address"
+      <*> o .: "value"
+      <*> o .: "datum"
+      <*> o .: "referenceScript"
 
 viewCompactTxOut ::
   forall era.